% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_report.R
\name{process_report}
\alias{process_report}
\title{Add together two numbers}
\usage{
process_report(
  report,
  task_description = "description",
  start_time = "start",
  end_time = "end",
  duration = "dur",
  client = "client",
  project = "project"
)
}
\arguments{
\item{report}{a dataframe of time tracking summary stats
from togglr API or website.}
}
\description{
Add together two numbers
}
