% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bar_chart.R
\name{time_bar_chart}
\alias{time_bar_chart}
\title{Make a bar plot showing percentage of time spent on each project}
\usage{
time_bar_chart(
  report,
  project = "project",
  project_time_percent = "project_time_percent"
)
}
\arguments{
\item{report}{a dataframe of time tracking summary stats
from togglr API or website.}
}
\description{
Make a bar plot showing percentage of time spent on each project
}
