% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/togglvis-package.R
\docType{package}
\name{togglvis-package}
\alias{togglvis}
\alias{togglvis-package}
\title{togglvis: Vizualize toggl data}
\description{
Simple functions to clean and visualize data recorded using the toggl app.
}
\keyword{internal}
